Minor maintenance fixes for derivativeDefintion.mlx
1) In the cooling coffee example, update the axis references so that
even if another plot is current when the code is run, the additional
axis ticks marking the point (t, T(t)) and adding the legend will occur
on the correct plot.

2) In the basketball example, correct the formula used in the original
creation of the plot to match the formula used in the updates of the
plot so the red dotted vertical line will properly track time in all
cases.

3) Update the function machine and derivative machine images to be
complete, including backgrounds, so they can be reused in situations
with different background colors without unexpected changes to the
design.